+++ packages/ragbits-document-search/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Fix reranker options typing (#562)
 - Add query rephraser options (#560)
 - Rename DocumentMeta create_text_document_from_literal to from_literal (#561)
 - Update audit imports (#427)